merapihkan baris code
