+++ src/db/db.js
@@ -1,0 +1,5 @@
+import { createClient } from "@supabase/supabase-js";
+import dotenv from "dotenv";
+dotenv.config();
+
+export const supabase = createClient(process.env.LINKDATABASE, process.env.APIKEY);
